refactor filter tasks
(cherry picked from commit 9e82039)
